# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem Mahad (Absensi, Hafalan, dll)",
        "title": "Mahad Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/hafalan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk mengambil data hafalan secara terpaginated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hafalan"
                ],
                "summary": "Mengambil semua data hafalan dengan pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Halaman yang ingin diambil",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Jumlah data per halaman",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hafalan fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch hafalan",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk menambahkan hafalan baru oleh mentor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hafalan"
                ],
                "summary": "Menambahkan hafalan baru",
                "parameters": [
                    {
                        "description": "Create Hafalan Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateHafalanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Hafalan created successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Mahasantri not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to create hafalan",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/hafalan/mahasantri/{mahasantri_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk mengambil data hafalan berdasarkan MahasantriID, dengan dukungan filtering berdasarkan kategori dan juz serta pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hafalan"
                ],
                "summary": "Mengambil semua hafalan berdasarkan MahasantriID dengan pagination dan filtering",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Mahasantri",
                        "name": "mahasantri_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "ziyadah",
                            "murojaah"
                        ],
                        "type": "string",
                        "description": "Filter by kategori",
                        "name": "kategori",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1, 2",
                        "description": "Filter by juz",
                        "name": "juz",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hafalan fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Mahasantri not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch hafalan",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/hafalan/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk mengambil data hafalan berdasarkan ID yang diberikan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hafalan"
                ],
                "summary": "Mendapatkan hafalan berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Hafalan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hafalan found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Hafalan not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch hafalan",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk memperbarui data hafalan yang sudah ada berdasarkan ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hafalan"
                ],
                "summary": "Memperbarui data hafalan berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Hafalan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data untuk memperbarui hafalan",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateHafalanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hafalan updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or no changes detected",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Hafalan not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to update hafalan",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk menghapus data hafalan berdasarkan ID yang diberikan.",
                "tags": [
                    "Hafalan"
                ],
                "summary": "Menghapus hafalan berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Hafalan",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hafalan deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Hafalan not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to delete hafalan",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/hafalan/{mahasantri_id}/kategori": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Endpoint ini digunakan untuk mengambil data hafalan berdasarkan MahasantriID dan kategori (ziyadah atau murojaah), dengan dukungan pagination.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hafalan"
                ],
                "summary": "Mengambil semua hafalan berdasarkan MahasantriID dan kategori dengan pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Mahasantri",
                        "name": "mahasantri_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "ziyadah",
                            "murojaah"
                        ],
                        "type": "string",
                        "description": "Kategori untuk memfilter hafalan",
                        "name": "kategori",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hafalan by category fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Mahasantri not found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch hafalan",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateHafalanRequest": {
            "type": "object",
            "required": [
                "halaman",
                "juz",
                "kategori",
                "mahasantri_id",
                "total_setoran",
                "waktu"
            ],
            "properties": {
                "catatan": {
                    "type": "string"
                },
                "halaman": {
                    "type": "string"
                },
                "juz": {
                    "type": "integer",
                    "maximum": 30,
                    "minimum": 1
                },
                "kategori": {
                    "type": "string",
                    "enum": [
                        "ziyadah",
                        "murojaah"
                    ]
                },
                "mahasantri_id": {
                    "type": "integer"
                },
                "total_setoran": {
                    "type": "number",
                    "minimum": 0
                },
                "waktu": {
                    "type": "string",
                    "enum": [
                        "shubuh",
                        "isya"
                    ]
                }
            }
        },
...
        "dto.UpdateHafalanRequest": {
            "type": "object",
            "properties": {
                "catatan": {
                    "type": "string"
                },
                "halaman": {
                    "type": "string"
                },
                "juz": {
                    "type": "integer"
                },
                "kategori": {
                    "type": "string"
                },
                "total_setoran": {
                    "type": "number"
                },
                "waktu": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}